scripts/requirements.txt
# Requirements for Cancer Gene GC Content Analysis Project
# Install with: pip install -r requirements.txt

biopython>=1.83
numpy>=1.26.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
scipy>=1.11.0
